requirements.txt
flask
google-generativeai
faiss-cpu
numpy
